Move get_pulp_additional_config out of conftest.py
It does not seem to belong there, especially considering that is is only
used in one test file.
